pyproject.toml
[tool.poetry]
name = "inspari_config"
version = "0.1.6"
description = "Inspari configuration utilities"
authors = ["emilhe <emil.h.eriksen@gmail.com>"]
readme = "README.md"
packages = [
    { include = "inspari" }
]

[tool.poetry.dependencies]
python = "^3.11"
azure-identity = "^1.19.0"
azure-keyvault = "^4.2.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
